flaskapp/requirements.txt
alembic==1.4.3
Babel==2.8.0
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-editor==1.0.4
python-slugify==4.0.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.20
text-unidecode==1.3
Werkzeug==1.0.1
WTForms==2.3.3
gunicorn  
